updates to the site that include an accordion menu, pre-loader animation and a page transition.
The accordion menu broke when I took a functional template I created and added it into the games.html. The functionality was dead in the water, but the formatting was good. The error in the inspect, shows an uncaught ReferenceError: $ which is undefined. I could not figure out how to resolve that successfully as it all appears as I would expect, but I also do not fully understand what I am looking at yet.